requirements.txt
discord.py>=2.3.0
openai>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pydub>=0.25.1
stripe>=7.0.0
flask>=3.0.0
ngrok>=0.0.1 